[REFACTOR] admin api를 위한 /admin prefix Controller 작성, 테스트 이관
